requirements.txt

---

### **`requirements.txt`**

```plaintext
Flask==2.2.3
Pillow==10.0.0
pytesseract==0.3.10
